+++ src/Console/Commands/IndexCommand.php
@@ -23,7 +23,7 @@ class IndexCommand extends Command
         $containers = Containers::enabled();
 
         if ($containers->isEmpty()) {
-            $this->components->error(__('asset-usage::messages.no_containers'));
+            $this->components->error('No asset containers are enabled for this addon.');
 
             return self::FAILURE;
         }